Generate type adapters for lists using XList
